requirements.txt
streamlit
cohere
pinecone-client
python-dotenv
PyPDF2
python-docx
tqdm
langchain

